Reduce code duplication using a common EEPROM storage adapter.
